feat(table-view): move "Query" button next to the Data/Schema toggle
Previously the "Query" button lived inside the DataGrid toolbar.
That meant it was only reachable while the Data tab was active —
on the Schema tab the toolbar isn't rendered, so users had to
flip back to Data just to spawn a query console for the same
table. Pinning it to the TableView header alongside the
Data/Schema toggle makes it a view-level affordance, equally
reachable from both modes.

Changes:

- TableView.tsx
  - New `handleOpenQuery` (identical semantics to the one that
    used to live in DataGrid; uses the `connectionColor` prop
    that's already plumbed in, so no new store dependency)
  - New `.tv-query-btn` rendered in the existing `tv-header-right`,
    immediately before the Data/Schema segmented control
- TableView.css
  - `.tv-query-btn` styled to match the segmented-control buttons
    (same 32px height, same font weight, same letter-spacing) so
    the right side of the header reads as one cohesive cluster
- DataGrid.tsx
  - Removed the now-redundant Query button + handler
  - Dropped the `Terminal` icon import (no remaining users in
    this file)

No behavioural changes for users beyond the relocation: same tab
spawn semantics, same `Query - <table>` title, same tooltip.

Verified: 630/630 frontend tests pass, `tsc --noEmit` clean.
